+++ lib/hexapdf/type/acro_form/text_field.rb
@@ -100,9 +100,24 @@ module HexaPDF
           unflag(:file_select, :multiline, :password)
         end
 
+        # Initializes the text field to be a password field.
+        #
+        # This method should only be called directly after creating a new text field because it
+        # doesn't completely reset the object.
+        def initialize_as_password_field
+          delete(:V)
+          flag(:password)
+          unflag(:comb, :multiline, :file_select)
+        end
+
         # Returns +true+ if this field is a comb text field.
         def comb_text_field?
           flagged?(:comb) && !(flagged?(:file_select) || flagged?(:multiline) || flagged?(:password))
+        end
+
+        # Returns +true+ if this field is a password field.
+        def password_field?
+          flagged?(:password) && !(flagged?(:file_select) || flagged?(:multiline) || flagged?(:comb))
         end
 
         # Returns the field value, i.e. the text contents of the field, or +nil+ if no value is set.