+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 * Support for creating AcroForm comb text fields and their appearances
+* Support for creating AcroForm password fields and their appearances
 
 
 ## 0.13.0 - 2020-11-15